+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog] and this project adheres to
 - `Integer` object type
   - Features `From` conversions for every native Rust integer type, including
     `u128` and `i128`
+  - Supports logical bitwise operations
   - Methods:
     - `pack` and `unpack` for converting to and from words respectively
     - `to_truncated` for converting similarly to `as` with primitives